add code for showing an artist not working and initial coding to show routing for countries in the browse by country. Link goes to the browse by country page
