+++ client/components/Greetings.js
@@ -8,10 +8,10 @@
          <h1>Welcome To Arm The Pit</h1>
          <div>
            <h2>Browse Artist By: </h2>
-           <Link to="/country"><button type="button" className="btn btn-info">COUNTRY</button></Link>
-           <Link to="/genre"><button type="button" className="btn btn-info">GENRE</button></Link> 
-           <Link to="/name"><button type="button" className="btn btn-info">NAME</button></Link> 
-           <Link to="/state"><button type="button" className="btn btn-info">STATE</button></Link>   
+           <Link to="/browsecountry"><button type="button" className="btn btn-info">COUNTRY</button></Link>
+           <Link to="/browsegenre"><button type="button" className="btn btn-info">GENRE</button></Link> 
+           <Link to="/browsename"><button type="button" className="btn btn-info">NAME</button></Link> 
+           <Link to="/browsestate"><button type="button" className="btn btn-info">STATE</button></Link>   
         </div>     
        </div>
      );
