+++ client/rootReducer.js
@@ -2,6 +2,7 @@ import { combineReducers } from 'redux';
  
  import flashMessages from './reducers/flashMessages';
  import auth from './reducers/auth';
+ import artist from './reducers/artist';
  import artists from './reducers/artists';
  import countries from './reducers/countries';
  import genres from './reducers/genres';
@@ -11,6 +12,7 @@ import { combineReducers } from 'redux';
  export default combineReducers({
    flashMessages,
    auth,
+   artist,
    artists,
    countries,
    genres,
