+++ public/index.html
@@ -13,6 +13,6 @@
 </head>
 <body>
 	<div id="app"></div>
-	<script src="bundle.js"></script>
+	<script src="/bundle.js"></script>
 </body>
 </html>
