+++ server/index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
 	<div id="app"></div>
-	<script src="bundle.js"></script>
+	<script src="/bundle.js"></script>
 </body>
 </html>
