+++ server/index.js
@@ -9,6 +9,7 @@ const artists = require('./routes/artists');
 const artistsAuth = require('./routes/auth/artists');
 const fans = require('./routes/fans');
 const fansAuth = require('./routes/auth/fans');
+const artist = require('./routes/artist/artist.js');
 const byArtist = require('./routes/browse/byartist');
 const byCountry = require('./routes/browse/bycountry');
 const byGenre = require('./routes/browse/bygenre');
@@ -27,6 +28,7 @@ app.use('/api/artists', artists);
 app.use('/api/auth/artists', artistsAuth);
 app.use('/api/fans', fans);
 app.use('/api/auth/fans', fansAuth);
+app.use('api/artist/', artist);
 app.use('/api/browse/byartist', byArtist);
 app.use('/api/browse/bycountry', byCountry);
 app.use('/api/browse/bygenre', byGenre);
